Crate firmware_core

This crate

This is the core part of the firmware for a 3D printer.

This crate is an Hardware Abstraction Layer (HAL), meaning it provides an implementation of:

Communication with the external world, through an HTTP server (over WiFi) for example

A basic file system to store and later retrieve G-code files from a flash memory

Parsing and execution of G-code

Closed loop PID heaters

A motion controller to move the tool that is connected to the stepper motors, with support for homing and bed leveling

Some drivers for various components (like BLTouch, TMC2209, fans…)

OTA updates

Pausing and resuming a print

But it doesn’t provide any platform specific implementation, which is done by other crates (like the esp32-s3 one).

This makes the adding of support for a new platform easy since only the specific hardware implementations need to be written. (which means implementing some traits from [embedded_hal] and some other custom traits present in this crate).

Re-exports

pub use embedded_hal;

Modules

printer

A module for managing a 3D printer’s components, communication, and operations.

utils

This module provides various utility functions and types that assist in the firmware’s functionality. It includes operations for mathematical calculations, measurements, and data conversions. The utilities are designed to simplify common tasks and improve code reusability across different components of the firmware.