Module firmware_core::printer::communication

This module contains everything related to making the 3D printer able to communicate with the external world. This could be thanks to:

WiFi

USB (not supported yet)

Bluetooth (not supported yet)

The data exchange with the external world is used to monitor the printer’s status, make it execute commands (like start printing a 3D model)…

Modules

communicator

http

ota

security

Structs

Communication

CommunicationConfig

MultiThreadCommunication

Enums

CreationError

SendablePeripherals

TickError