firmware_core::printer::communication
Module communicator

This module handles the communication between the microcontroller and the external world.

It provides the Communicator trait that defines methods for registering HTTP request handlers and managing communication protocols such as WiFi and USB.

The communication is essential for:

Receiving commands from external sources

Sending status updates

Handling various types of requests defined in the firmware

Modules

wifi

This module provides the functionality for WiFi communication in the microcontroller.

Traits

Communicator

A type that implements this trait can be used to make the microcontroller communicate with the external world.