firmware_core::printer::communication::communicator
Trait Communicator

pub trait Communicator {
    type Error: Debug;

    // Required method
    fn register_request<P: Peripherals + 'static>(
        &mut self,
        request: HttpRequest,
        resources: Resources<P>,
    ) -> Result<(), Self::Error>;

    // Provided method
    fn register_all_requests<P: Peripherals + 'static>(
        &mut self,
        resources: Resources<P>,
    ) -> Result<(), Self::Error> { ... }
}

A type that implements this trait can be used to make the microcontroller communicate with the external world.

Required Associated Types

type Error: Debug

The type of error that can occur during communication.

Required Methods

fn register_request<P: Peripherals + 'static>( &mut self, request: HttpRequest, resources: Resources<P>, ) -> Result<(), Self::Error>

Registers a handler for a specific HTTP request.

This method allows for individual request handling to be defined.

Arguments

request - The specific HTTP request to register.

resources - Resources to be used when handling the request.

Returns

A result indicating success or failure.

Provided Methods

fn register_all_requests<P: Peripherals + 'static>( &mut self, resources: Resources<P>, ) -> Result<(), Self::Error>

Registers handlers for all possible HTTP requests supported by this firmware.

This method iterates through all defined HttpRequest types and registers each one with the provided resources.

Arguments

resources - Resources to be used when handling requests.

Returns

A result indicating success or failure.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.