firmware_core::printer::communication
Enum SendablePeripherals

pub enum SendablePeripherals<P: Peripherals> {
    ComponentsThread {

        watchdog_creator: Option<P::WatchdogCreator>,
        stepper_ticker_timer: Option<P::StepperTickerTimer>,
        kinematics: Option<P::Kinematics>,
        left_motor_dir_pin: Option<P::LeftDirPin>,
        left_motor_step_pin: Option<P::LeftStepPin>,
        right_motor_dir_pin: Option<P::RightDirPin>,
        right_motor_step_pin: Option<P::RightStepPin>,
        z_axis_motor_dir_pin: Option<P::ZAxisDirPin>,
        z_axis_motor_step_pin: Option<P::ZAxisStepPin>,
        extruder_motor_dir_pin: Option<P::ExtruderDirPin>,
        extruder_motor_step_pin: Option<P::ExtruderStepPin>,
        uart_driver: Option<P::UartDriver>,
        x_axis_endstop: Option<P::XAxisEndstop>,
        y_axis_endstop: Option<P::YAxisEndstop>,
        z_axis_endstop: Option<P::ZAxisEndstop>,
        layer_fan_pin: Option<P::FanPin>,
        hotend_fan_pin: Option<P::FanPin>,
        bed_cartridge_heater_pin: Option<P::HeatedBedHeaterPin>,
        bed_thermistor_pin: Option<P::HeatedBedAdcPin>,
        hotend_cartridge_heater_pin: Option<P::CartridgeHeaterPin>,
        hotend_thermistor_pin: Option<P::HotendAdcPin>,
        adc: Option<P::Adc>,
        system_time: Option<P::SystemTime>,
    
},
    CommunicationThread {
        watchdog_creator: Option<P::WatchdogCreator>,
        system_time: Option<P::SystemTime>,
        flash_chip: Option<P::FlashChip>,
        flash_spi: Option<P::FlashSpi>,
        wifi_driver: Option<P::WifiDriver>,
        server: Option<Box<dyn FnOnce() -> Result<P::Server, P::ServerError> + Send>>,
        ota: Option<P::Ota>,
    },
}

Represents peripherals that can be sent between threads.

This enum allows for safe transmission of peripheral data across thread boundaries in the context of communication and components management.

Variants

ComponentsThread

Contains peripherals required for the components thread.

Fields

watchdog_creator: Option<P::WatchdogCreator>

stepper_ticker_timer: Option<P::StepperTickerTimer>

kinematics: Option<P::Kinematics>

left_motor_dir_pin: Option<P::LeftDirPin>

Motor control pins for various axes.

left_motor_step_pin: Option<P::LeftStepPin>

right_motor_dir_pin: Option<P::RightDirPin>

right_motor_step_pin: Option<P::RightStepPin>

z_axis_motor_dir_pin: Option<P::ZAxisDirPin>

z_axis_motor_step_pin: Option<P::ZAxisStepPin>

extruder_motor_dir_pin: Option<P::ExtruderDirPin>

extruder_motor_step_pin: Option<P::ExtruderStepPin>

uart_driver: Option<P::UartDriver>

UART driver for serial communication.

x_axis_endstop: Option<P::XAxisEndstop>

Endstop sensors for the axes.

y_axis_endstop: Option<P::YAxisEndstop>

z_axis_endstop: Option<P::ZAxisEndstop>

layer_fan_pin: Option<P::FanPin>

Fan control pins.

hotend_fan_pin: Option<P::FanPin>

bed_cartridge_heater_pin: Option<P::HeatedBedHeaterPin>

Heater and thermistor pins for the bed and hotend.

bed_thermistor_pin: Option<P::HeatedBedAdcPin>

hotend_cartridge_heater_pin: Option<P::CartridgeHeaterPin>

hotend_thermistor_pin: Option<P::HotendAdcPin>

adc: Option<P::Adc>

ADC for reading sensor values.

system_time: Option<P::SystemTime>

System time for managing delays.

CommunicationThread

Contains peripherals required for the communication thread.

Fields

watchdog_creator: Option<P::WatchdogCreator>

system_time: Option<P::SystemTime>

flash_chip: Option<P::FlashChip>

Flash chip for memory operations.

flash_spi: Option<P::FlashSpi>

SPI interface for the flash chip.

wifi_driver: Option<P::WifiDriver>

WiFi driver for connectivity.

server: Option<Box<dyn FnOnce() -> Result<P::Server, P::ServerError> + Send>>

HTTP server for handling requests.

ota: Option<P::Ota>

OTA (Over-the-Air) update interface.

Implementations

impl<P: Peripherals> SendablePeripherals<P>

pub fn of_communication_thread(peripherals: &mut P) -> Self

Constructs a SendablePeripherals instance for the communication thread.

Arguments

peripherals - Mutable reference to the peripherals from which data will be taken.

Returns

A SendablePeripherals instance configured for the communication thread.

pub fn of_components_thread(peripherals: &mut P) -> Self

Constructs a SendablePeripherals instance for the components thread.

Arguments

peripherals - Mutable reference to the peripherals from which data will be taken.

Returns

A SendablePeripherals instance configured for the components thread.