firmware_core::printer::communication::http::other::printer_state
Function tick

pub fn tick<P: Peripherals>(
    hotend_pid_controller: &TemperaturePidController<P::CartridgeHeaterPin, P::Adc, P::HotendAdcPin>,
    bed_pid_controller: &TemperaturePidController<P::HeatedBedHeaterPin, P::Adc, P::HeatedBedAdcPin>,
)

Updates the “screenshot” of the state of some components of the machine based on the current and target temperatures of the two PID controllers (saves the result in the static instance of PrinterState so that you can later retrieve the state using get_current_state).