Enum firmware_core::printer::communication::http::request::HttpRequest

pub enum HttpRequest {
    ListFiles,
    DeleteFile,
    PrintFile,
    SendFile,
    GetPrintStatus,
    PauseOrResume,
    PrinterState,
    Move,
    ListGCodeCommandsInMemory,
    SendGCodeCommands,
    OTAUpdate,
}

A possible request that the HTTP server in this firmware can handle. Each request has a method, an URI and a callback function that handles it.

Variants

ListFiles

List the metadatas of all the G-code files saved in the file system.

DeleteFile

Delete a specific file from the file system.

PrintFile

Start printing a specific file.

SendFile

Send a G-code file to the printer (that later on could be printed).

GetPrintStatus

Get the status of the current print (if a print is in execution), providing some info like the expected duration, the name of the file being printed…

PauseOrResume

Pause or resume (based on the previous state) the current print.

PrinterState

Get the status of various components of the machine (like the current temperature of the hotend, or the target temperature of the bed).

Move

Moves the tool of the machine by the specified amount on the possible directions.

ListGCodeCommandsInMemory

Provide the history of G-code commands that have been read from the file system to print the current file. This request must be issued often, since microcontrollers don’t have much RAM (which means that the history being tracked can’t be long), and a large amount of time between different requests might lead to some G-code commands that have been executed by the printer to not appear in the history.

SendGCodeCommands

“Manually” add the provided G-code commands to the command buffer of the current print process.

OTAUpdate

Over-The-Air update.

Implementations

impl HttpRequest

pub fn get_method(&self) -> Method

Returns the HTTP method that the request received by the server should have to invoke the callback of this variant of the enum.

pub fn get_uri(&self) -> &'static str

Returns the relative URI that the request received by the server should have to invoke the callback of this variant of the enum.

pub fn get_callback<C: Connection, P: Peripherals>( &self ) -> fn(_: Request<&mut C>, _: Resources<P>) -> HandlerResult

Returns a pointer to the function that should be called when the server receives a request with the corresponding URI and method.