firmware_core::printer::components::drivers::fan
Struct Fan

pub struct Fan<P: PwmPin> { /* private fields */ }

A fan connected to the microcontroller that can be controlled using the P pin.

Implementations

impl<P: PwmPin> Fan<P>

pub fn new(pin: P, minimum_duty_cycle_fan_moves: Percentage) -> Self

Returns a Fan that can control its speed through the provided pin, and starts moving when the speed set is greater or equal to minimum_duty_cycle_fan_moves.

Warning

The speed that you’ll later set using Self::set_speed will be remapped to the moveable range. For example this:

let mut fan = Fan::new(pin, Percentage::from_0_to_1(0.5).unwrap());
fan.set_speed(Percentage::from_0_to_1(0.8).unwrap());

will internally set a duty cycle of 90%, but since the fan starts moving at a 50% duty cycle, it will effectively move at the 80% of its maximum speed.

pub fn set_speed( &mut self, speed: Percentage, ) -> Result<(), <P as PwmPin>::Error>

Sets the speed percentage at which the fan rotates.

Returns Ok(()) if the speed was set correctly, otherwise returns Err(error).

Check Self::new for more info.