firmware_core::printer::components::drivers
Module servo_motor

This module provides the implementation for controlling a servo motor using PWM signals.

The ServoMotor struct allows for positioning a servo motor by sending pulse durations on a specified pin. The ServoPosition struct represents the angle at which the servo should be set.

Structs

ServoMotor

A servo motor connected to the microcontroller that can be controlled by sending pulses on the P pin.

ServoPosition

Position held by a ServoMotor.

Constants

REFRESH_RATE

The 50Hz value usually used by servo motors.