Module firmware_core::printer::components::drivers::spi_flash_memory

Structs

MT29F2G01ABAGDWB

2Gbit 3.3V NAND SPI flash memory chip (datasheet).

SpiFlashMemory

A flash memory connected to the microcontroller through a SPI interface.

Enums

FeatureRegister

The register of a feature in the MT29F2G01ABAGDWB chip.

ValidateIdError

An error returned from SpiFlashMemory::validate_id.

Traits

FlashMemoryChip

A type that represents a flash memory chip.

FlashMemoryChipExt

Extra functionality provided automatically to every type that implements FlashMemoryChip.