firmware_core::printer::components::drivers::spi_flash_memory
Struct SpiFlashMemory

pub struct SpiFlashMemory<Chip: FlashMemoryChip, Spi: SpiDevice<u8>> { /* private fields */ }

A flash memory connected to the microcontroller through a SPI interface.

Warning

It’s really important to understand how a flash memory works before you use any of the method in this struct, or you might have unexpected results and/or wear out the memory.

Implementations

impl<Chip: FlashMemoryChip, Spi: SpiDevice<u8>> SpiFlashMemory<Chip, Spi>

pub fn new(spi: Spi, chip: Chip) -> Self

pub fn program( &mut self, data: &[u8], address: u32, ) -> Result<(), <Spi as ErrorType>::Error>

Program the provided data in the flash memory starting from the provided address, writing to data.len() / Chip::PAGE_SIZE sequential pages.

Returns Ok(()) if all the bytes have been written successfully, otherwise returns Err(...).

Note

The program will only affect the data area of a page, not the ECC one.

Warning

You can’t update some data in the memory you programmed before without erasing it first.

Check struct's documentation.

pub fn read( &mut self, address: u32, data: &mut [u8], ) -> Result<(), <Spi as ErrorType>::Error>

Reads [data.len()] bytes from the data areas of the pages starting from the specified address.

Returns Ok(()) if all the bytes have been read successfully, otherwise returns Err(...).

Note

The read will only affect the data area of a page, not the ECC one.

pub fn read_ecc( &mut self, row_address: RowAddress<Chip>, data: &mut [u8], ) -> Result<(), <Spi as ErrorType>::Error>

Reads [data.len()] bytes from the ECC area of the page identified by the provided row_address.

Returns Ok(()) if all the bytes have been read successfully, otherwise returns Err(...).

Note

The read will only affect the ECC area of a page, not the data one.

Panics

Panics if data.len() > Chip::PAGE_ECC_SIZE.

pub fn internal_data_move( &mut self, from: RangeInclusive<u32>, to_start_address: u32, ) -> Result<(), <Spi as ErrorType>::Error>

Moves the bytes in the from address range to the address range of the same size that starts at to_start_address.

It’s an internal move because bytes aren’t copied in the microcontroller and then programmed at the right location, but everything is done in the flash memory chip.

Warning

Check struct's documentation.

pub fn reset(&mut self) -> Result<(), <Spi as ErrorType>::Error>

Puts the flash memory in a known condition.

Warning

Calling this function can block the microcontroller for some milliseconds.

pub fn erase_blocks( &mut self, block_indices_to_erase: RangeInclusive<u16>, ) -> Result<(), <Spi as ErrorType>::Error>

Erases all the blocks in the provided block_indices_to_erase range by resetting all the bits of those blocks to 1.

Returns Ok(()) if all the blocks have been reset successfully, otherwise returns Err(...).

Warning

Check struct's documentation.

pub fn validate_id(&mut self) -> Result<(), ValidateIdError<Spi>>

Validate that the IDs (ManufacturerID and DeviceID) of the connected flash memory chip are the same as the ones written in the FlashMemoryChip trait.

Returns Ok(()) if the microcontroller was able to communicate with the flash memory and the returned IDs are correct, otherwise returns Err(ValidateIdError).

pub fn get_features( &mut self, features: FeatureRegister, ) -> Result<u8, <Spi as ErrorType>::Error>

Returns the value stored in the provided features register of the flash memory chip.

Check the datasheet of the chip to understand what each feature does.

pub fn set_features( &mut self, features: FeatureRegister, features_value: u8, ) -> Result<(), <Spi as ErrorType>::Error>

Sets the value stored in the provided features register of the flash memory chip to features_value.

Check the datasheet of the chip to understand what each feature does.

pub fn get_chip_mut(&mut self) -> &mut Chip

Returns a mutable reference to the underlying chip.