Trait firmware_core::printer::components::drivers::spi_flash_memory::FlashMemoryChipExt

pub trait FlashMemoryChipExt: FlashMemoryChip {
    const BLOCK_SIZE: u32 = _;
    const PLANE_SIZE: u32 = _;
    const LUN_SIZE: u32 = _;
    const MEMORY_SIZE: u32 = _;

    // Provided methods
    fn get_address_of_block_index(block_index: u16) -> u32 { ... }
    fn get_block_index_of_address(address: u32) -> u16 { ... }
}

Extra functionality provided automatically to every type that implements FlashMemoryChip.

Provided Associated Constants

const BLOCK_SIZE: u32 = _

Size of the data area of a block.

const PLANE_SIZE: u32 = _

Size of the data area of a plane.

const LUN_SIZE: u32 = _

Size of the data area of a LUN.

const MEMORY_SIZE: u32 = _

Size of the data area of the chip.

Provided Methods

fn get_address_of_block_index(block_index: u16) -> u32

Returns the memory address of the first byte of the first page of the block at the provided block_index.

fn get_block_index_of_address(address: u32) -> u16

Returns the index of the block that contains the provided address.

Object Safety

This trait is not object safe.