Enum firmware_core::printer::components::drivers::stepper_motor::tmc2209::MicrostepsPerStep

#[repr(u8)]
pub enum MicrostepsPerStep {
    FULLSTEP = 0,
    M2 = 1,
    M4 = 2,
    M8 = 3,
    M16 = 4,
    M32 = 5,
    M64 = 6,
    M128 = 7,
    M256 = 8,
}

Amount of microsteps the motor will require to do a full single step.

Variants

FULLSTEP = 0

1 microstep corresponds to a step.

M2 = 1

2 microsteps correspond to a step.

M4 = 2

4 microsteps correspond to a step.

M8 = 3

8 microsteps correspond to a step.

M16 = 4

16 microsteps correspond to a step.

M32 = 5

32 microsteps correspond to a step.

M64 = 6

64 microsteps correspond to a step.

M128 = 7

128 microsteps correspond to a step.

M256 = 8

256 microsteps correspond to a step.

Implementations

impl MicrostepsPerStep

pub const MAX_RESOLUTION: Self = Self::M256

Maximum microstep resolution achievable by the TMC2209 chip.

assert_eq!(MicrostepsPerStep::MAX_RESOLUTION, MicrostepsPerStep::M256);

pub const fn as_exponent_of_2(&self) -> u8

pub const fn as_value(&self) -> u16

Returns the number of microsteps per step this instance represents.

Examples

assert_eq!(MicrostepsPerStep::FULLSTEP.as_value(), 1);
assert_eq!(MicrostepsPerStep::M2.as_value(), 2);
assert_eq!(MicrostepsPerStep::M4.as_value(), 4);
assert_eq!(MicrostepsPerStep::M8.as_value(), 8);
assert_eq!(MicrostepsPerStep::M16.as_value(), 16);
assert_eq!(MicrostepsPerStep::M32.as_value(), 32);
assert_eq!(MicrostepsPerStep::M64.as_value(), 64);
assert_eq!(MicrostepsPerStep::M128.as_value(), 128);
assert_eq!(MicrostepsPerStep::M256.as_value(), 256);

pub const fn as_max_resolution_microsteps_count(&self) -> u16

Returns the number of microsteps with the highest resolution this instance represents.

Examples

assert_eq!(MicrostepsPerStep::FULLSTEP.as_max_resolution_microsteps_count(), 256);
assert_eq!(MicrostepsPerStep::M2.as_max_resolution_microsteps_count(), 128);
assert_eq!(MicrostepsPerStep::M4.as_max_resolution_microsteps_count(), 64);
assert_eq!(MicrostepsPerStep::M8.as_max_resolution_microsteps_count(), 32);
assert_eq!(MicrostepsPerStep::M16.as_max_resolution_microsteps_count(), 16);
assert_eq!(MicrostepsPerStep::M32.as_max_resolution_microsteps_count(), 8);
assert_eq!(MicrostepsPerStep::M64.as_max_resolution_microsteps_count(), 4);
assert_eq!(MicrostepsPerStep::M128.as_max_resolution_microsteps_count(), 2);
assert_eq!(MicrostepsPerStep::M256.as_max_resolution_microsteps_count(), 1);