firmware_core::printer::components::drivers::stepper_motor::tmc2209
Struct TMC2209

pub struct TMC2209 { /* private fields */ }

A driver to communicate with the TMC2209 chip (to control a stepper motor).

Only UART communication is supported.

This struct contains various methods to change the settings of the TMC2209 chip writing to its internal registers. Keep in mind that talking to the TMC2209 chip will block the microcontroller for as few as some hundreds of microseconds to as much as variuos milliseconds (it depends on the baud rate of the UART driver).

For further information on what each setting will do check TMC2209's datasheet.

Implementations

impl TMC2209

pub const STEP_MIN_LOW_TIME: SmallDuration = _

pub const STEP_MIN_HIGH_TIME: SmallDuration = _

pub fn new_using_uart<Uart: UartTrait>( uart_address: UARTAddress, uart_driver: &mut Uart, microsteps_per_step: MicrostepsPerStep, ) -> Result<Self, Uart::Error>

Returns a TMC2209 driver that is connected controlled by the microcontroller using UART.

This will also set some default settings on the TMC2209 that should be good for a 3D printer. The driver is disabled by default, so call TMC2209::enable when you want the motor to start being able to rotate.

pub fn set_enabled<Uart: UartTrait>( &mut self, enabled: bool, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

pub fn enable<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Enable the TMC2209 via software (UART).

This switches on the power stage on the TMC2209, making the motor able to move.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn disable<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Disable the TMC2209 via software (UART).

This switches off the power stage on the TMC2209, making all motor outputs floating (the motor can’t move).

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn enable_stealth_chop<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Enable StealthChop on the driver, disabling SpreadCycle.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn disable_stealth_chop<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Disable StealthChop on the driver, enabling SpreadCycle.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn set_microsteps_per_step<Uart: UartTrait>( &mut self, microsteps_per_step: MicrostepsPerStep, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Sets the amount of microsteps the motor will take for each single step.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn get_microsteps_per_step(&self) -> MicrostepsPerStep

Get the value you previously set using Self::set_microsteps_per_step.

pub fn read_pin_state<Uart: UartTrait>( &self, register_field: IOIN, uart_driver: &mut Uart, ) -> Result<u8, ReadRegisterError<Uart>>

Read the state of a physical pin of the TMC2209 chip using UART.

Blocking

Since the value is read from the TMC2209’s registers, calling this function will block the microcontroller for a bit.

pub fn enable_automatic_current_scaling<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Enable automatic current control.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn disable_automatic_current_scaling<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Disable automatic current control.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn enable_automatic_gradient_adaptation<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Makes the TMC2209 chip automatically adjust the PWM gradient value.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn disable_automatic_gradient_adaptation<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Stops the automatic adjustment of the PWM gradient value in the TMC2209 chip.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn set_current_scaling<Uart: UartTrait>( &mut self, on_run: Option<Percentage>, on_hold: Option<Percentage>, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Scales the current going to the motors.

on_run is the scale used when the motor is running, on_hold instead when it is idle.

Pass None as a parameter to not change the scale you had before.

Examples

let mut uart_driver = //...
let mut tmc2209 = //...

// This will set the current scaling of the motor when running to 80% and leave
// the current scaling when the motor is idle unchanged.
tmc2209.set_current_scaling(Some(Percentage::from_0_to_1(0.8).unwrap()), None, &mut uart_driver);

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn enable_analog_scale_current<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn disable_analog_scale_current<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn enable_microstep_interpolation<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

The TMC2209 chip will extrapolate the microsteps per step you previously set to 256.

This won’t make any difference in the way you control the motor, it will just make the motor less noisy at the cost of a small positional error.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn disable_microstep_interpolation<Uart: UartTrait>( &mut self, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>

Disable the microstep interpolation. Check TMC2209::enable_microstep_interpolation for more info.

Blocking

Since the value is sent to the TMC2209’s chip, calling this function will block the microcontroller for a bit.

pub fn set_reply_delay<Uart: UartTrait>( &mut self, delay: ReplyDelay, uart_driver: &mut Uart, ) -> Result<(), Uart::Error>