firmware_core::printer::components::file_system::regions::metadata
Struct FilesMetadatasRegion

pub struct FilesMetadatasRegion { /* private fields */ }

Implementations

impl FilesMetadatasRegion

pub fn read_from_flash<Chip: FlashMemoryChip, Spi: SpiDevice<u8>>( spi_flash_memory: &mut SpiFlashMemory<Chip, Spi>, regions_config: &RegionsConfig, ) -> Result<Self, <Spi as ErrorType>::Error>

Reads the metadatas region you previously stored from the provided spi_flash_memory at the address specified in regions_config.metadata_block_range or creates a default one if it’s the first time the flash memory is used.

Returns Ok(Self) if the region was correctly read or if it was missing and the default one was succesfully created and automatically stored in the chip. Otherwise returns Err(...) (if there has been an error in communicating with the spi_flash_memory).

pub fn store_in_flash<Chip: FlashMemoryChip, Spi: SpiDevice<u8>>( &self, spi_flash_memory: &mut SpiFlashMemory<Chip, Spi>, regions_config: &RegionsConfig, ) -> Result<(), <Spi as ErrorType>::Error>

Stores this region in the provided spi_flash_memory at the address specified in regions_config.metadata_block_range so that you can later recover it (even after shutting down the microcontrooler) using Self::read_from_flash.

Returns Err(...) if there has been an error in communicating with the flash memory, otherwise returns Ok(()).

pub fn create_file<Chip: FlashMemoryChip>( &mut self, bytes_count: u32, regions_config: &RegionsConfig, ) -> Result<(FileId, u32), NotEnoughSpaceAvailable>

Finds a space large enough to store data_size bytes contiguously.

Returns Ok((new_file_id, new_file_start_address)) if the space has been found, otherwise returns Err(NotEnoughSpaceAvailable)

Warning

This won’t store the metadata of the newly created file in the flash memory, a call to Self::finish_writing_file is required.

pub fn start_writing_file<Chip: FlashMemoryChip, Spi: SpiDevice<u8>>( &mut self, file_metadata: FileMetadata, spi_flash_memory: &mut SpiFlashMemory<Chip, Spi>, regions_config: &RegionsConfig, ) -> Result<(), <Spi as ErrorType>::Error>

pub fn finish_writing_file<Chip: FlashMemoryChip, Spi: SpiDevice<u8>>( &mut self, file_id: FileId, spi_flash_memory: &mut SpiFlashMemory<Chip, Spi>, regions_config: &RegionsConfig, ) -> Result<(), <Spi as ErrorType>::Error>

pub fn get_file_metadata(&self, file_id: FileId) -> Option<FileMetadata>

If a file with the specified file_id is stored in this region, returns its metadata wrapped in Some. Otherwise returns None.

pub fn does_file_exist(&self, file_id: FileId) -> bool

Returns true if a file with the specified file_id is stored in this region, otherwise returns false.

pub fn delete_file( &mut self, file_id: FileId, ) -> Result<FileMetadata, FileDoesntExist>

Removes the file with the specified file_id from the region.

Warning

A call to Self::store_in_flash is required after calling this method to actually store the modified region in the flash memory.

pub fn get_files_metadatas(&self) -> &[FileMetadata]

Returns a reference to the metadatas of all the files stored in this region.

pub fn get_file_validator(&self) -> FileMetadataValidator

Returns the current file validator.

pub fn compact<Chip: FlashMemoryChip, Spi: SpiDevice<u8>>( &mut self, spi_flash_memory: &mut SpiFlashMemory<Chip, Spi>, ) -> Result<(), <Spi as ErrorType>::Error>

Collect all the fragmented data holes present in the memory in one big chunk, moving all the files next to each other.

Check this for more info.