firmware_core::printer::components
Module g_code

G-code is the language understood by 3D printers, and it’s used to make the 3D printer execute commands (like move the tool to a specific location, set the bed’s temperature…).

A 3D printer slicer software receives a 3D model and converts it to a long list of G-code commands, which are then sent to the printer so that it can print the model.

This firmware (for now) only supports the most important G-code commands.

For an extensive list of all the existing G-code commands, check the RepRap’s documentation.

Modules

commands

A list of all the G-code commands supported by this firmware.

execute

parameters

G-code commands can have 0, 1, or more parameters. Each parameter starts with an identifier (like X or Y) and then there is the value of the parameter (which can be a string, a bool, a float…)

parser

This module contains utilities to convert strings (which could be read from the file system) to executable G-code commands.

Enums

Status

Status of execution (or preparation) of a G-code command.

Traits

AsAny

This trait is only used for testing purposes, and it’s automatically implemented by every GCodeCommand.

GCodeCommand