firmware_core::printer::components::g_code::execute
Struct GCodeExecuter

pub struct GCodeExecuter<P: Peripherals> { /* private fields */ }

Implementations

impl<P: Peripherals> GCodeExecuter<P>

pub fn tick( &mut self, printer_components: &mut Printer3DComponents<P>, ) -> Result<(), TickError>

pub fn has_command_to_execute(&self) -> bool

Returns true if it is currently executing a command or it has at least 1 command in its command buffer. Otherwise returns false.

pub fn set_units(&mut self, units: Units)

pub fn set_position_mode(&mut self, mode: PositionMode)

pub fn set_extruder_position_mode(&mut self, mode: PositionMode)

pub fn calculate_position_based_on_mode( &self, printer_components: &Printer3DComponents<P>, x: &mut Option<Distance>, y: &mut Option<Distance>, z: &mut Option<Distance>, e: &mut Option<Distance>, )

pub fn add_command_to_buffer(&mut self, command: Box<dyn GCodeCommand<P>>)

pub fn add_commands_to_buffer( &mut self, commands: Vec<Box<dyn GCodeCommand<P>>>, )

pub fn save_position( &mut self, position: VectorN<4>, slot: usize, ) -> Result<(), InvalidPositionSlot>

Save the provided position at the specified slot, so that you can later retrieve it using [Self::get_position(slot)].

Returns Err(InvalidPositionSlot) if slot >= SAVED_POSITIONS_COUNT.

Examples

let mut g_code_executer = GCodeExecuter::<MockPeripherals>::default();

assert!(g_code_executer.save_position(VectorN::default(), 0).is_ok());
assert!(g_code_executer.save_position(VectorN::default(), 1000).is_err());

pub fn get_saved_position( &self, slot: usize, ) -> Result<VectorN<4>, InvalidPositionSlot>

Get the position you provided to Self::save_position at the same slot as the one you provided here (or VectorN::default if you never saved a position in that slot).

Returns Err(InvalidPositionSlot) if slot >= SAVED_POSITIONS_COUNT.

Examples

let mut g_code_executer = GCodeExecuter::<MockPeripherals>::default();

assert_eq!(g_code_executer.get_saved_position(0), Ok(VectorN::default()));

g_code_executer.save_position(VectorN::new([Distance::from_centimeters(3); 4]), 0).unwrap();
assert_eq!(g_code_executer.get_saved_position(0), Ok(VectorN::new([Distance::from_centimeters(3); 4])));

assert_eq!(g_code_executer.get_saved_position(100000), Err(InvalidPositionSlot));