firmware_core::printer::components::g_code
Module parameters

G-code commands can have 0, 1, or more parameters. Each parameter starts with an identifier (like X or Y) and then there is the value of the parameter (which can be a string, a bool, a float…)

Examples

X20.1

Y-5.3

S60

Modules

identifier

A list of all the identifiers in the parameters of a command supported by this firmware.

value

A list of all the values in the parameters of a command supported by this firmware.

Structs

Param

A parameter of a G-code command.