Trait firmware_core::printer::components::g_code::GCodeCommand

pub trait GCodeCommand<P: Peripherals>: Send + Sync + Debug + AsAny<P> {
    // Provided methods
    fn prepare(
        &mut self,
        _: &mut Printer3DComponents<P>,
        _: &mut GCodeExecuter<P>
    ) -> Status { ... }
    fn execute(
        &mut self,
        _: &mut Printer3DComponents<P>,
        _: &mut GCodeExecuter<P>
    ) -> Status { ... }
}

Provided Methods

fn prepare( &mut self, _: &mut Printer3DComponents<P>, _: &mut GCodeExecuter<P> ) -> Status

This method is called as soon as the G-code command is parsed from a string and before it is put in the command buffer of the GCodeExecuter. When this method returns Status::Finished, the command is inserted in the buffer.

fn execute( &mut self, _: &mut Printer3DComponents<P>, _: &mut GCodeExecuter<P> ) -> Status

This method is called when the G-code command becomes the first command in the command buffer of the GCodeExecuter. When this method returns Status::Finished, the command is removed from the buffer.