firmware_core::printer::components::hal::adc
Trait AdcPinExt

pub trait AdcPinExt<A: Adc>: AdcPin<A> {
    // Required method
    fn read_percentage(
        &mut self,
        adc: &mut A,
    ) -> Result<Percentage, ReadPercentageError<A, Self>>
       where Self: Sized;
}

Extension trait for ADC pins that provides additional functionalities (it is automatically implemented).

Required Methods

fn read_percentage( &mut self, adc: &mut A, ) -> Result<Percentage, ReadPercentageError<A, Self>>
where Self: Sized,

Reads a value from the ADC pin and returns it as a percentage.

This method uses the ADC’s maximum readable value to convert the raw reading into a percentage. It returns an error if the reading cannot be performed or if the result is not a valid percentage.