firmware_core::printer::components::hal::interrupt
Trait InterruptPin

pub trait InterruptPin {
    type Error: Debug;

    // Required method
    unsafe fn subscribe_to_interrupt(
        &mut self,
        when_to_trigger: Trigger,
        callback: impl FnMut() + Send + 'static,
    ) -> Result<(), Self::Error>;
}

Trait representing an interrupt-capable pin.

Required Associated Types

type Error: Debug

The type of error that can occur when subscribing to an interrupt.

Required Methods

unsafe fn subscribe_to_interrupt( &mut self, when_to_trigger: Trigger, callback: impl FnMut() + Send + 'static, ) -> Result<(), Self::Error>

Subscribes to an interrupt on this pin with a specified trigger condition.

Safety

The callback will be called in an Interrupt Service Routine (ISR) context.

This means that the callback should be lightweight and should not perform operations that could block or take a long time. Care must also be taken to ensure that shared resources are accessed in a thread-safe manner.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.