firmware_core::printer::components::hal::timer
Trait Timer

pub trait Timer {
    type Error: Debug;
    type AdditionalFunctionality: TimerAdditionalFunctionality + Send;

    // Required methods
    fn get_additional_functionality(&self) -> Self::AdditionalFunctionality;
    fn get_clock_frequency(&self) -> Frequency;
    unsafe fn on_alarm(
        &mut self,
        callback: impl FnMut() + Send + 'static,
    ) -> Result<(), Self::Error>;
    fn enable_alarm(&mut self, enable: bool) -> Result<(), Self::Error>;
    fn get_alarm_in_ticks(&self) -> Result<u64, Self::Error>;

    // Provided method
    fn get_alarm(&self) -> Result<Duration, Self::Error> { ... }
}

A timer that can keep track of elapsed time and call a provided callback when a certain time is reached.

Required Associated Types

type Error: Debug

The type of error that can occur when using the timer.

type AdditionalFunctionality: TimerAdditionalFunctionality + Send

Type for additional timer functionality.

Required Methods

fn get_additional_functionality(&self) -> Self::AdditionalFunctionality

Returns an instance of Self::AdditionalFunctionality for accessing extended timer functionalities, like setting alarm times or retrieving the current time.

The separation into two traits allows for calling these methods within the callback provided to Timer::on_alarm, which would be impossible if they were in the same trait.

fn get_clock_frequency(&self) -> Frequency

Returns the frequency at which the clock of the timer is running.

unsafe fn on_alarm( &mut self, callback: impl FnMut() + Send + 'static, ) -> Result<(), Self::Error>

Calls the provided callback every time the alarm time set using TimerAdditionalFunctionality::set_alarm is reached.

Safety

The callback will be called in an ISR context.

fn enable_alarm(&mut self, enable: bool) -> Result<(), Self::Error>

Enable or disable the timer based on the provided enable variable.

When the timer is disabled, it will not keep track of time, and alarms will not fire.

fn get_alarm_in_ticks(&self) -> Result<u64, Self::Error>

Retrieves the current alarm value in ticks.

Provided Methods

fn get_alarm(&self) -> Result<Duration, Self::Error>

Retrieves the current alarm as a Duration.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.