firmware_core::printer::components::hal::timer
Trait TimerAdditionalFunctionality

pub trait TimerAdditionalFunctionality: 'static {
    type Error: Debug;

    // Required methods
    fn set_alarm(&mut self, time: Duration) -> Result<(), Self::Error>;
    fn set_alarm_in_ticks(&mut self, ticks: u64) -> Result<(), Self::Error>;
    fn get_time(&self) -> Result<Duration, Self::Error>;
    fn get_time_in_ticks(&self) -> Result<u64, Self::Error>;
}

Trait for additional timer functionality.

Required Associated Types

type Error: Debug

The type of error that can occur in the additional functionality.

Required Methods

fn set_alarm(&mut self, time: Duration) -> Result<(), Self::Error>

Sets an alarm that calls the provided callback when the current time reaches the specified time.

fn set_alarm_in_ticks(&mut self, ticks: u64) -> Result<(), Self::Error>

Sets an alarm in ticks.

fn get_time(&self) -> Result<Duration, Self::Error>

Gets the current time kept by the timer.

fn get_time_in_ticks(&self) -> Result<u64, Self::Error>

Gets the current time in ticks.