firmware_core::printer::components::hal::watchdog
Trait Watchdog

pub trait Watchdog {
    type Error: Debug;

    // Required method
    fn feed(&mut self) -> Result<(), Self::Error>;
}

A type that can restart the watchdog timer in the microcontroller by calling Self::feed.

Required Associated Types

type Error: Debug

The type of error that can occur while feeding the watchdog.

Required Methods

fn feed(&mut self) -> Result<(), Self::Error>

Feeds the watchdog to reset the timer and prevent the firmware from entering an error state.

Check the trait's documentation for more information.