firmware_core::printer::components::hal::watchdog
Trait WatchdogCreator

pub trait WatchdogCreator {
    type Watchdog: Watchdog;

    // Required method
    fn watch_current_thread(self) -> Option<Self::Watchdog>;
}

A type that can instantiate a Watchdog on the current thread by calling the Self::watch_current_thread method.

After you watch a thread, you must regularly feed the returned watchdog; otherwise, the firmware will enter an error state.

Required Associated Types

type Watchdog: Watchdog

The type of watchdog that this creator produces.

Required Methods

fn watch_current_thread(self) -> Option<Self::Watchdog>

Creates a new watchdog for the current thread.

Check the trait's documentation for more information.