Module firmware_core::printer::components::motion

Modules

axes

Helpers to deal with the axes of the machine.

bed_leveling

homing

kinematics

planner

Translates movements (for example a 10mm move in the X axis) to motion profiles “understood” by stepper motors.

ticker

Structs

CreationConfig

CreationParameters

Data necessary to create a MotionController.

MotionController

Controls all the stuff (moves, homing, bed levelling..) related to the movement of everything inside the machine.

MotorConfig

Data necessary to configure a stepper motor.

RotationToLinearMotion

An helper struct used to calculate the distance taken by an object connected to a mechanism that converts the rotation of a motor to a linear displacement of the object.

Enums

CreationError

An error that can occur when you instatiate a MotionController struct.

SetPausedError

An error that can occur when you pause or resume a MotionController.

Constants

N_MOTORS

Number of stepper motors controlled by the machine.