firmware_core::printer::components::motion::homing
Module endstop

Submodule for handling endstop functionality, which is critical for detecting limits and homing operations in motion control systems.

Structs

ManualEndstop

A “fake” Endstop that is not physically present in the machine, but it has to be triggered via software calling ManualEndstop::trigger.

Traits

Endstop

Type that can be used as an endstop.