firmware_core::printer::components::motion::homing::endstop
Struct ManualEndstop

pub struct ManualEndstop { /* private fields */ }

A “fake” Endstop that is not physically present in the machine, but it has to be triggered via software calling ManualEndstop::trigger.

This could be useful for testing purposes.

Examples

let mut endstop = ManualEndstop::new();
let triggered = Arc::new(AtomicBool::new(false));
let triggered_clone = Arc::clone(&triggered);
endstop.trigger();
// Nothing happens because no callback has been registered yet
assert_eq!(triggered.load(Ordering::Relaxed), false);

unsafe { endstop.on_end_reached(move || triggered_clone.store(true, Ordering::Relaxed)) };
endstop.trigger();

assert_eq!(triggered.load(Ordering::Relaxed), true);

Implementations

impl ManualEndstop

pub fn new() -> Self

Returns a ManualEndstop without any callback registered.

To set the callback function you should use ManualEndstop::on_end_reached, and to trigger the callback use ManualEndstop::trigger.

Check the struct's doc for an example.

pub fn trigger(&mut self)

Sets the value returned by Self::is_end_reached to true.

If a callback has been registered, it will be called.

Check the struct's doc for an example.

pub fn remove_trigger(&mut self)

Sets the value returned by Self::is_end_reached to false.

Examples

let mut endstop = ManualEndstop::new();

assert_eq!(endstop.is_end_reached(), Ok(false));

// This won't call any callback since there's no callback registered but it will set the `is_end_reached` return value to true.
endstop.trigger();
assert_eq!(endstop.is_end_reached(), Ok(true));

endstop.remove_trigger();
assert_eq!(endstop.is_end_reached(), Ok(false));