firmware_core::printer::components::motion
Module kinematics

A module for different kinematics implementations, such as Cartesian and CoreXY.

Structs

CartesianKinematics

“Cartesian” kinematics.

CoreXYKinematics

CoreXY kinematics.

Traits

Kinematics

Type that represents the kinematics of the machine (i.e. how the motion of the a and b motors effects the movement of the tool carriage).