firmware_core::printer::components::motion
Module planner

Translates movements (for example a 10mm move in the X axis) to motion profiles “understood” by stepper motors.

The generated motion profiles are optimized to make the move as fast as possible without provoking problems (for example making a stepper lose steps).

Credits

This module incorporates code from the Marlin firmware project, which is licensed under the GNU General Public License (GPL). Portions of this module are derived from Marlin, for more information about Marlin and its licensing terms, please visit https://github.com/MarlinFirmware/Marlin.

Disclaimer

This firmware project is not affiliated with or endorsed by the Marlin firmware project. While portions of this module are derived from Marlin and are subject to the GNU General Public License (GPL), this project is independently developed and maintained by Angelo Cipriani. For official information and support regarding the Marlin firmware, please refer to the Marlin project’s documentation and community resources.

Modules

communicate_to_ticker

Let the main thread communicate with the ticker ISR.

Structs

Block

Result of a planned move. It contains all the data necessary to allow the StepperMotorsTicker to make the move actually happen.

BlocksBufferIsFull

MoveId

Planner

The planning involves calculating the most optimal speed and acceleration profiles

Settings

Enums

Flag

Type Aliases

StepsPerSecond