Module firmware_core::printer::components::motion::planner::communicate_to_ticker

Structs

Communication

Data shared between the “main thread” and the ticker ISR.

KinematicsFunctions

Function pointers used to convert the motion of the a and b stepper motors to movements of the tool in the X and Y axes (check set_kinematics_functions).

Functions

get_blocks

Tries to lock the internal Mutex and returns Some(Communication) if it has been successful, otherwise returns None.

is_block_available

Returns true if there is either the current or the next block is Some in get_blocks.

is_homing_block_available

Returns true if either the current or the next block have the Flag::Homing set.

print_logs_from_isr

set_kinematics_functions

Sets the KinematicsFunctions used by the StepperMotorsTicker to understand how the rotation of some motors relates to the movement of the tool.

store_blocks_if_necessary

Checks if the ticker ISR requires 1 or 2 blocks and eventually uses them.