Struct firmware_core::printer::components::motion::planner::Planner

pub struct Planner<const N: usize> { /* private fields */ }

The planning involves calculating the most optimal speed and acceleration profiles

Implementations

impl Planner<N_MOTORS>

pub fn tick(&mut self) -> Option<[i32; 4]>

impl<const N: usize> Planner<N>

pub fn new( blocks_count: usize, stepper_motors_ticker: &StepperMotorsTicker<impl Timer>, settings: Settings<N> ) -> Self

pub fn get_settings_mut(&mut self) -> &mut Settings<N>

Returns a mutable reference to the Settings you provided to Self::new.

pub fn get_position(&self) -> &VectorN<N>

Returns the target_position of the last planned move.

pub fn set_position(&mut self, position: VectorN<N>)

Manually sets the current position of the planner, making it think that the next planned move will start from this position.

pub fn mark_last_added_move_as_ready_to_go(&mut self)

pub fn has_move_been_executed(&self, move_to_check: MoveId) -> bool

pub fn plan_move<K: Kinematics>( &mut self, target_position: VectorN<N>, steps_per_mm: [f32; N], move_speed_mm_s: f32 ) -> Result<MoveId, BlocksBufferIsFull>

Plans a move that will make the tool move to the target_position starting from the target_position of the last planned move.

steps_per_mm contains the number of steps required to move the tool by 1mm on each axis.

The planned move won’t be sent to the StepperMotorsTicker until you call Self::mark_last_added_move_as_ready_to_go.

Check Self for more info.

pub fn has_any_move_planned(&self) -> bool

pub fn has_any_move_with_set_flags_planned(&self, flags: EnumSet<Flag>) -> bool

pub fn set_flags_on_last_added_block(&mut self, flags: EnumSet<Flag>)