Struct firmware_core::printer::components::motion::MotionController

pub struct MotionController<Timer: TimerTrait, Kinematics: KinematicsTrait, ZEndstop: ZAxisProbe> { /* private fields */ }

Controls all the stuff (moves, homing, bed levelling..) related to the movement of everything inside the machine.

Implementations

impl<Timer: TimerTrait, Kinematics: KinematicsTrait, ZEndstop: ZAxisProbe> MotionController<Timer, Kinematics, ZEndstop>

pub fn new<LeftDirPin: OutputPin + Send + 'static, LeftStepPin: OutputPin + Send + 'static, RightDirPin: OutputPin + Send + 'static, RightStepPin: OutputPin + Send + 'static, ZAxisDirPin: OutputPin + Send + 'static, ZAxisStepPin: OutputPin + Send + 'static, ExtruderDirPin: OutputPin + Send + 'static, ExtruderStepPin: OutputPin + Send + 'static, XEndstop: Endstop + Send + 'static, YEndstop: Endstop + Send + 'static, Uart: UartTrait>( peripherals: CreationParameters<Timer, Kinematics, LeftDirPin, LeftStepPin, RightDirPin, RightStepPin, ZAxisDirPin, ZAxisStepPin, ExtruderDirPin, ExtruderStepPin, XEndstop, YEndstop, ZEndstop>, configuration: CreationConfig, uart_driver: &mut Uart ) -> Result<Self, CreationError<Timer, ZEndstop, Uart>>

pub fn plan_move( &mut self, x: Option<Distance>, y: Option<Distance>, z: Option<Distance>, e: Option<Distance>, feed_rate: Option<f32> ) -> Result<MoveId, BlocksBufferIsFull>

Make the tool move to the provided coordinates after the previous moves are completed (and after Self::mark_last_move_as_ready_to_go has been called!).

x, y, z, e (extruder) are the movements along each respective axis, and for the ones of them that are set to None there won’t be any movement along that axis.

The optional feed_rate will determine the speed of not only this move, but also all the subsequent ones.

Returns Err(BlocksBufferIsFull) if the move couldn’t be planned, and you MUST call this method again to try to plan it!

Warning

This motion controller must be ticked to effectively execute the moves.

pub fn mark_last_move_as_ready_to_go(&mut self)

Make the last planned move ready to be executed.

pub fn has_move_been_executed(&self, move_to_check: MoveId) -> bool

pub fn get_last_planned_move_end_position(&self) -> Option<VectorN<N_MOTORS>>

Returns the position of the last planned move if there has been one, otherwise returns None.

pub fn set_position( &mut self, x: Option<Distance>, y: Option<Distance>, z: Option<Distance>, e: Option<Distance> )

pub fn tick(&mut self) -> Result<(), TickError<Probe<ZEndstop>>>

This method internally ticks the HomingProcedure and the Planner, executing the planned moves.

pub fn set_paused(&mut self, paused: bool) -> Result<(), SetPausedError<Timer>>

pub fn get_current_position_esteem(&self) -> Option<VectorN<N_MOTORS>>

pub fn start_homing(&mut self) -> Result<(), BlocksBufferIsFull>

Make the machine start the HomingProcedure after all the planned moves are completed.

Returns Err(BlocksBufferIsFull) if the procedure couldn’t be started, and you MUST call this method again to try to start it!

pub fn start_bed_leveling(&mut self)

pub fn get_tmc2209_driver(&mut self, axis: Axis) -> &mut TMC2209

Returns a mutable reference to the TMC2209 driver you provided to Self::new that drives the stepper motor on the specific axis.

Warning

It’s better if you don’t modify any setting of any TMC2209 driver while there is a print going on. This is because talking to the chip needs some time, and even delays are used in internal TMC2209 functions. A delay could block the StepperMotorsTicker from doing its job resulting in missed steps and bad prints.

Consider changing all the TMC2209’s settings before or after a print.