Struct firmware_core::printer::components::motion::ticker::StepperMotorsTicker

pub struct StepperMotorsTicker<Timer: TimerTrait> { /* private fields */ }

Attaches an interrupt to the timer you provide to Self::new that checks if stepper motors should take a step based on the planned block communicated to the ticker, and eventually makes them take the step.

Implementations

impl<Timer: TimerTrait> StepperMotorsTicker<Timer>

pub fn new<LeftDirPin: OutputPin + Send + 'static, LeftStepPin: OutputPin + Send + 'static, RightDirPin: OutputPin + Send + 'static, RightStepPin: OutputPin + Send + 'static, ZAxisDirPin: OutputPin + Send + 'static, ZAxisStepPin: OutputPin + Send + 'static, ExtruderDirPin: OutputPin + Send + 'static, ExtruderStepPin: OutputPin + Send + 'static, XEndstop: Endstop + Send + 'static, YEndstop: Endstop + Send + 'static, ZEndstop: ZAxisProbe>( left_motor: StepperMotor<LeftDirPin, LeftStepPin>, right_motor: StepperMotor<RightDirPin, RightStepPin>, z_axis_motor: StepperMotor<ZAxisDirPin, ZAxisStepPin>, extruder_motor: StepperMotor<ExtruderDirPin, ExtruderStepPin>, timer: Timer, x_endstop: XEndstop, y_endstop: YEndstop, z_endstop: &mut Probe<ZEndstop> ) -> Result<Self, CreationError<Timer, ZEndstop>>

Returns Err(Timer::Error) if there was an error while setting the callback of the timer, otherwise in case of success returns Ok(Self).

Warning

You must call Self::enable after creating this instance to actually make the ticker work.

pub fn enable(&mut self) -> Result<(), EnableError<Timer>>

Enable the inner timer so that the ticker can make the stepper motors move (by pulsing the STEP pin at the right moment and with the right value on the DIR pin).

pub fn disable(&mut self) -> Result<(), <Timer as TimerTrait>::Error>

Disable the inner timer, stopping this ticker from making the stepper motors move.

pub fn get_tick_frequency(&self) -> Frequency

Get the frequency of the inner timer. This represents the minimum interval of time between two steps taken by a stepper motor.

pub fn get_time( &self ) -> Result<Duration, <Timer::AdditionalFunctionality as TimerAdditionalFunctionality>::Error>