firmware_core::printer::components::pauser
Function is_paused

pub fn is_paused() -> bool

Returns false if you’ve never called any function in this module (as a default value) or if after calling them (resume and toggle_pause) you end up with false being the current state.

Check the module's documentation for more details.

Examples

assert_eq!(pauser::is_paused(), false);

pauser::resume();
assert_eq!(pauser::is_paused(), false);

pauser::resume();
assert_eq!(pauser::is_paused(), false);

pauser::toggle_pause();
assert_eq!(pauser::is_paused(), true);