firmware_core::printer::components
Module print_process

This module manages the process of printing files by parsing G-code commands.

It provides the PrintProcess struct, which is responsible for controlling the printing workflow, including reading files from the file system, parsing G-code commands, and executing them. It also handles the buffer management for G-code commands and ensures that the print process progresses correctly.

Structs

LineToExecuteParsed

Contains the parsed result of a line intended for execution.

PrintProcess

This struct controls the process of printing a file, by parsing the content of the file to G-code commmands.

PrintProcessOk

The call to PrintProcess::tick has been successful, and this struct contains the string that has been read from the flash memory (Self::read_lines) and also the result of parsing that string to GCodeCommands (in Self::read_commands).

Enums

PrintProcessError

The call to PrintProcess::tick hasn’t been successful. This enum contains the problems that could have arised.

Functions

get_commands_in_buffer_count

Gets the current count of commands in the buffer.

set_commands_in_buffer_count

Sets the number of commands currently in the buffer.