Struct firmware_core::printer::components::print_process::PrintProcess

pub struct PrintProcess<P: Peripherals> { /* private fields */ }

This struct controls the process of printing a file, by parsing the content of the file to G-code commmands.

Implementations

impl<P: Peripherals> PrintProcess<P>

pub fn new(max_commands_in_buffer_before_reading_new: u16) -> Self

Returns an empty PrintProcess.

Use Self::print_file to start the process of printing.

pub fn print_file( &mut self, file_id_to_print: FileId, current_time: Option<Duration> )

Starts printing the file with the provided file_id_to_print file id.

Warning

You must call Self::tick to effectively make the print process progress.

pub fn tick( &mut self, file_system: &mut FileSystem<P::FlashChip, P::FlashSpi>, commands_in_buffer: u16 ) -> Result<PrintProcessOk<P>, PrintProcessError<P::FlashSpi>>

If a file is currently being printed, calling this function will try to read new G-code commands from the file system that will be executed by the GCodeExecuter.

If instead no file is being printed, calling this function will do nothing.

pub fn get_file_being_printed(&self) -> Option<FileId>

Returns Some(FileId) if a file is currently being printed (which means you called Self::print_file and the file has not been completely read yet), otherwise returns None.

pub fn get_print_estimated_duration_in_seconds(&self) -> Option<u32>

Returns Some(duration_in_secs) if a file is currently being printed (which means you called Self::print_file and the file has not been completely read yet) and in the file there’s a line containing ;TIME: {value} where {value} is a number. Otherwise returns None.

pub fn get_print_start_time(&self) -> Option<Duration>

Returns Some(print_start_time) if a file is currently being printed (which means you called Self::print_file and the file has not been completely read yet) and you provided a start time for the print. Otherwise returns None.

pub fn parse_line_to_execute<'a>( &mut self, line: &'a str ) -> Result<LineToExecuteParsed<'a, P>, ()>

Returns the command and the comment present in the provided line (if they are present), or Err(()) if the line is a GCodeLine::Error.