firmware_core::printer::components
Struct Printer3DComponents

pub struct Printer3DComponents<P: Peripherals> {
    pub layer_fan: Fan<P::FanPin>,
    pub hotend_fan: Fan<P::FanPin>,
    pub motion_controller: MotionController<P::StepperTickerTimer, P::Kinematics, P::ZAxisEndstop>,
    pub uart_driver: P::UartDriver,
    pub hotend_pid_controller: TemperaturePidController<P::CartridgeHeaterPin, P::Adc, P::HotendAdcPin>,
    pub heated_bed_pid_controller: TemperaturePidController<P::HeatedBedHeaterPin, P::Adc, P::HeatedBedAdcPin>,
    pub adc: P::Adc,
    pub clock: Clock<P::SystemTime>,
    pub g_code_executer: Option<GCodeExecuter<P>>,
}

This struct encapsulates all the elements required to make a 3D print possible, including fans, motion control, temperature controllers, and the G-code executer.

It manages the initialization and operation of these components.

Fields
layer_fan: Fan<P::FanPin>hotend_fan: Fan<P::FanPin>motion_controller: MotionController<P::StepperTickerTimer, P::Kinematics, P::ZAxisEndstop>uart_driver: P::UartDriverhotend_pid_controller: TemperaturePidController<P::CartridgeHeaterPin, P::Adc, P::HotendAdcPin>heated_bed_pid_controller: TemperaturePidController<P::HeatedBedHeaterPin, P::Adc, P::HeatedBedAdcPin>adc: P::Adcclock: Clock<P::SystemTime>g_code_executer: Option<GCodeExecuter<P>>
Implementations

impl<P: Peripherals> Printer3DComponents<P>

pub fn new( peripherals: &mut P, config: ComponentsConfig, ) -> Result<Self, CreationError<P::StepperTickerTimer, P::ZAxisEndstop, P::UartDriver>>

Creates a new instance of Printer3DComponents.

This method initializes the printer components using the provided peripherals and configuration settings. It ensures all necessary components are properly set up.

Arguments

peripherals - The hardware peripherals required for the printer.

config - Configuration settings for the printer’s components.

Returns

A Result containing the initialized Printer3DComponents instance or a CreationError if an error occurs.

pub fn tick( &mut self, ) -> Result<(), TickError<P::ZAxisEndstop, P::UartDriver, P::StepperTickerTimer>>

Updates the state of the printer components, performing necessary tasks.

This method should be called periodically to ensure the printer operates correctly. It manages tasks related to temperature control, motion, and G-code execution.