firmware_core::printer::components::temperature::safety
Module allowed_range

Defines mechanisms for checking if a temperature is within an allowed range.

This module provides the AllowedTemperatureRangeSafety struct, which ensures that a given temperature is within a specified safe range. It is useful for enforcing safety constraints in temperature control systems.

Structs

AllowedTemperatureRangeSafety

Makes sure a Temperature is in the allowed range you passed when constructing an instance of this struct.