firmware_core::printer::components::time
Trait SystemTime

pub trait SystemTime {
    // Required methods
    fn now(&self) -> Duration;
    fn delay(&self, duration: Duration);
}

A clock in the microcontroller that can keep time.

Required Methods

fn now(&self) -> Duration

Returns the amount of time this microcontroller has been running for.

fn delay(&self, duration: Duration)

Blocks this core of the microcontroller for the provided duration.