firmware_core::printer::components
Trait Peripherals

pub trait Peripherals {

    type WatchdogCreator: WatchdogCreator;
    type Kinematics: Kinematics;
    type StepperTickerTimer: Timer;
    type LeftDirPin: OutputPin + Send + 'static;
    type LeftStepPin: OutputPin + Send + 'static;
    type RightDirPin: OutputPin + Send + 'static;
    type RightStepPin: OutputPin + Send + 'static;
    type ZAxisDirPin: OutputPin + Send + 'static;
    type ZAxisStepPin: OutputPin + Send + 'static;
    type ExtruderDirPin: OutputPin + Send + 'static;
    type ExtruderStepPin: OutputPin + Send + 'static;
    type UartDriver: Uart;
    type XAxisEndstop: Endstop + Send + 'static;
    type YAxisEndstop: Endstop + Send + 'static;
    type ZAxisEndstop: ZAxisProbe + 'static;
    type FlashChip: FlashMemoryChip + Send + 'static;
    type FlashSpi: SpiDevice<u8> + Send + 'static;
    type CartridgeHeaterPin: PwmPin;
    type HotendAdcPin: AdcPin<Self::Adc>;
    type HeatedBedHeaterPin: PwmPin;
    type HeatedBedAdcPin: AdcPin<Self::Adc>;
    type Adc: Adc;
    type FanPin: PwmPin;
    type SystemTime: SystemTime + Send;
    type WifiDriver: Wifi + Send + 'static;
    type Server: HttpServer + 'static;
    type ServerError: Debug;
    type Ota: Ota + Send + 'static;

    // Required methods
    fn take_watchdog_creator(&mut self) -> Option<Self::WatchdogCreator>;
    fn take_kinematics(&mut self) -> Option<Self::Kinematics>;
    fn take_stepper_ticker_timer(&mut self) -> Option<Self::StepperTickerTimer>;
    fn take_left_motor_dir_pin(&mut self) -> Option<Self::LeftDirPin>;
    fn take_left_motor_step_pin(&mut self) -> Option<Self::LeftStepPin>;
    fn take_right_motor_dir_pin(&mut self) -> Option<Self::RightDirPin>;
    fn take_right_motor_step_pin(&mut self) -> Option<Self::RightStepPin>;
    fn take_z_axis_motor_dir_pin(&mut self) -> Option<Self::ZAxisDirPin>;
    fn take_z_axis_motor_step_pin(&mut self) -> Option<Self::ZAxisStepPin>;
    fn take_extruder_motor_dir_pin(&mut self) -> Option<Self::ExtruderDirPin>;
    fn take_extruder_motor_step_pin(&mut self) -> Option<Self::ExtruderStepPin>;
    fn take_uart_driver(&mut self) -> Option<Self::UartDriver>;
    fn take_x_axis_endstop(&mut self) -> Option<Self::XAxisEndstop>;
    fn take_y_axis_endstop(&mut self) -> Option<Self::YAxisEndstop>;
    fn take_z_axis_endstop(&mut self) -> Option<Self::ZAxisEndstop>;
    fn take_flash_chip(&mut self) -> Option<Self::FlashChip>;
    fn take_flash_spi(&mut self) -> Option<Self::FlashSpi>;
    fn take_adc(&mut self) -> Option<Self::Adc>;
    fn take_bed_thermistor_pin(&mut self) -> Option<Self::HeatedBedAdcPin>;
    fn take_bed_cartridge_heater_pin(
        &mut self,
    ) -> Option<Self::HeatedBedHeaterPin>;
    fn take_hotend_thermistor_pin(&mut self) -> Option<Self::HotendAdcPin>;
    fn take_hotend_cartridge_heater_pin(
        &mut self,
    ) -> Option<Self::CartridgeHeaterPin>;
    fn take_layer_fan_pin(&mut self) -> Option<Self::FanPin>;
    fn take_hotend_fan_pin(&mut self) -> Option<Self::FanPin>;
    fn take_system_time(&mut self) -> Option<Self::SystemTime>;
    fn take_wifi_driver(&mut self) -> Option<Self::WifiDriver>;
    fn get_ip_address_from_wifi_driver_function(    ) -> fn(_: &Self::WifiDriver) -> Option<IpAddr>;
    fn take_http_server(
        &mut self,
    ) -> Option<Box<dyn FnOnce() -> Result<Self::Server, Self::ServerError> + Send>>;
    fn take_ota(&mut self) -> Option<Self::Ota>;
    fn reboot_fn() -> fn();

}

A type that implements this trait contains all the peripherals required for the 3D printer’s operation.

You must implement this trait for a struct of the specific microcontroller you want to use with this firmware.

Required Associated Types

type WatchdogCreator: WatchdogCreator

A type implementing a watchdog timer creator.

type Kinematics: Kinematics

A type representing the kinematic model of the printer.

type StepperTickerTimer: Timer

A timer used for controlling stepper motors.

type LeftDirPin: OutputPin + Send + 'static

The output pin used to control the left motor’s direction.

type LeftStepPin: OutputPin + Send + 'static

The output pin used to control the left motor’s stepping.

type RightDirPin: OutputPin + Send + 'static

The output pin used to control the right motor’s direction.

type RightStepPin: OutputPin + Send + 'static

The output pin used to control the right motor’s stepping.

type ZAxisDirPin: OutputPin + Send + 'static

The output pin used to control the Z-axis motor’s direction.

type ZAxisStepPin: OutputPin + Send + 'static

The output pin used to control the Z-axis motor’s stepping.

type ExtruderDirPin: OutputPin + Send + 'static

The output pin used to control the extruder motor’s direction.

type ExtruderStepPin: OutputPin + Send + 'static

The output pin used to control the extruder motor’s stepping.

type UartDriver: Uart

A type for handling UART communication.

type XAxisEndstop: Endstop + Send + 'static

The type for the X-axis endstop.

type YAxisEndstop: Endstop + Send + 'static

The type for the Y-axis endstop.

type ZAxisEndstop: ZAxisProbe + 'static

The type for the Z-axis endstop.

type FlashChip: FlashMemoryChip + Send + 'static

A type representing a flash memory chip.

type FlashSpi: SpiDevice<u8> + Send + 'static

A type for interfacing with a SPI flash memory chip.

type CartridgeHeaterPin: PwmPin

A type representing the PWM pin for the cartridge heater.

type HotendAdcPin: AdcPin<Self::Adc>

A type for reading the hotend thermistor value.

type HeatedBedHeaterPin: PwmPin

A type representing the PWM pin for the heated bed.

type HeatedBedAdcPin: AdcPin<Self::Adc>

A type for reading the heated bed thermistor value.

type Adc: Adc

A type for handling analog-to-digital conversion.

type FanPin: PwmPin

A type representing the PWM pin for the fan.

type SystemTime: SystemTime + Send

A type for system time management.

type WifiDriver: Wifi + Send + 'static

A type for handling WiFi connections.

type Server: HttpServer + 'static

A type for managing an HTTP server.

type ServerError: Debug

A type for representing errors from the server.

type Ota: Ota + Send + 'static

A type for handling over-the-air updates.

Required Methods

fn take_watchdog_creator(&mut self) -> Option<Self::WatchdogCreator>

Attempts to take the watchdog creator peripheral. Returns None if the peripheral is not available.

fn take_kinematics(&mut self) -> Option<Self::Kinematics>

Attempts to take the kinematics peripheral. Returns None if the peripheral is not available.

fn take_stepper_ticker_timer(&mut self) -> Option<Self::StepperTickerTimer>

Attempts to take the stepper ticker timer peripheral. Returns None if the peripheral is not available.

fn take_left_motor_dir_pin(&mut self) -> Option<Self::LeftDirPin>

Attempts to take the left motor direction pin. Returns None if the peripheral is not available.

fn take_left_motor_step_pin(&mut self) -> Option<Self::LeftStepPin>

Attempts to take the left motor step pin. Returns None if the peripheral is not available.

fn take_right_motor_dir_pin(&mut self) -> Option<Self::RightDirPin>

Attempts to take the right motor direction pin. Returns None if the peripheral is not available.

fn take_right_motor_step_pin(&mut self) -> Option<Self::RightStepPin>

Attempts to take the right motor step pin. Returns None if the peripheral is not available.

fn take_z_axis_motor_dir_pin(&mut self) -> Option<Self::ZAxisDirPin>

Attempts to take the Z-axis motor direction pin. Returns None if the peripheral is not available.

fn take_z_axis_motor_step_pin(&mut self) -> Option<Self::ZAxisStepPin>

Attempts to take the Z-axis motor step pin. Returns None if the peripheral is not available.

fn take_extruder_motor_dir_pin(&mut self) -> Option<Self::ExtruderDirPin>

Attempts to take the extruder motor direction pin. Returns None if the peripheral is not available.

fn take_extruder_motor_step_pin(&mut self) -> Option<Self::ExtruderStepPin>

Attempts to take the extruder motor step pin. Returns None if the peripheral is not available.

fn take_uart_driver(&mut self) -> Option<Self::UartDriver>

Attempts to take the UART driver. Returns None if the peripheral is not available.

fn take_x_axis_endstop(&mut self) -> Option<Self::XAxisEndstop>

Attempts to take the X-axis endstop. Returns None if the peripheral is not available.

fn take_y_axis_endstop(&mut self) -> Option<Self::YAxisEndstop>

Attempts to take the Y-axis endstop. Returns None if the peripheral is not available.

fn take_z_axis_endstop(&mut self) -> Option<Self::ZAxisEndstop>

Attempts to take the Z-axis endstop. Returns None if the peripheral is not available.

fn take_flash_chip(&mut self) -> Option<Self::FlashChip>

Attempts to take the flash chip peripheral. Returns None if the peripheral is not available.

fn take_flash_spi(&mut self) -> Option<Self::FlashSpi>

Attempts to take the SPI flash peripheral. Returns None if the peripheral is not available.

fn take_adc(&mut self) -> Option<Self::Adc>

Attempts to take the ADC peripheral. Returns None if the peripheral is not available.

fn take_bed_thermistor_pin(&mut self) -> Option<Self::HeatedBedAdcPin>

Attempts to take the heated bed thermistor pin. Returns None if the peripheral is not available.

fn take_bed_cartridge_heater_pin(&mut self) -> Option<Self::HeatedBedHeaterPin>

Attempts to take the heated bed cartridge heater pin. Returns None if the peripheral is not available.

fn take_hotend_thermistor_pin(&mut self) -> Option<Self::HotendAdcPin>

Attempts to take the hotend thermistor pin. Returns None if the peripheral is not available.

fn take_hotend_cartridge_heater_pin( &mut self, ) -> Option<Self::CartridgeHeaterPin>

Attempts to take the hotend cartridge heater pin. Returns None if the peripheral is not available.

fn take_layer_fan_pin(&mut self) -> Option<Self::FanPin>

Attempts to take the layer fan pin. Returns None if the peripheral is not available.

fn take_hotend_fan_pin(&mut self) -> Option<Self::FanPin>

Attempts to take the hotend fan pin. Returns None if the peripheral is not available.

fn take_system_time(&mut self) -> Option<Self::SystemTime>

Attempts to take the system time peripheral. Returns None if the peripheral is not available.

fn take_wifi_driver(&mut self) -> Option<Self::WifiDriver>

Attempts to take the WiFi driver. Returns None if the peripheral is not available.

fn get_ip_address_from_wifi_driver_function() -> fn(_: &Self::WifiDriver) -> Option<IpAddr>

A function pointer to retrieve the IP address from the WiFi driver.

fn take_http_server( &mut self, ) -> Option<Box<dyn FnOnce() -> Result<Self::Server, Self::ServerError> + Send>>

Attempts to take the HTTP server. Returns None if the server is not available.

fn take_ota(&mut self) -> Option<Self::Ota>

Attempts to take the OTA update handler. Returns None if the peripheral is not available.

fn reboot_fn() -> fn()

Returns a function pointer that reboots the microcontroller.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.