firmware_core::utils::measurement::distance
Struct Distance

pub struct Distance { /* private fields */ }

A distance with a 10 nanometer sensitivity (10^-8 meters) and a range of values that goes from 10 * (-2^31) nm to 10 * (2^31-1) nm (which is from -21 to 21 meters).

Implementations

impl Distance

pub const CENTIMETER: Self = _

assert_eq!(Distance::CENTIMETER, Distance::from_centimeters(1));

pub const MILLIMETER: Self = _

assert_eq!(Distance::MILLIMETER, Distance::from_millimeters(1));

pub const MICROMETER: Self = _

assert_eq!(Distance::MICROMETER, Distance::from_micrometers(1));

pub const TENS_OF_NANOMETERS: Self = _

assert_eq!(Distance::TENS_OF_NANOMETERS, Distance::from_tens_of_nanometers(1));

pub const INCH: Self = _

assert_eq!(Distance::INCH, Distance::from_inches(1.));

pub const ZERO: Self = _

assert_eq!(Distance::ZERO, Distance::from_tens_of_nanometers(0));

pub const fn from_tens_of_nanometers(tens_of_nanometers: i32) -> Self

Returns a Distance from the provided tens of nanometers (10^-8 meters).

Examples

assert_eq!(Distance::from_tens_of_nanometers(30).as_tens_of_nanometers(), 30);
assert_eq!(Distance::from_tens_of_nanometers(1_400).as_tens_of_nanometers(), 1_400);
assert_eq!(Distance::from_tens_of_nanometers(510).as_micrometers(), 5);

pub const fn from_micrometers(micrometers: i32) -> Self

Returns a Distance from the provided micrometers (10^-6 meters).

Examples

assert_eq!(Distance::from_micrometers(30).as_micrometers(), 30);
assert_eq!(Distance::from_micrometers(1_400).as_micrometers(), 1_400);
assert_eq!(Distance::from_micrometers(5_100).as_millimeters(), 5);

pub const fn from_millimeters(millimeters: i32) -> Self

Returns a Distance from the provided millimeters (10^-3 meters).

Examples

assert_eq!(Distance::from_millimeters(30).as_millimeters(), 30);
assert_eq!(Distance::from_millimeters(1_400).as_millimeters(), 1_400);
assert_eq!(Distance::from_millimeters(2_000).as_centimeters(), 200);

pub const fn from_centimeters(centimeters: i32) -> Self

Returns a Distance from the provided centimeters (10^-2 meters).

Examples

assert_eq!(Distance::from_centimeters(30).as_centimeters(), 30);
assert_eq!(Distance::from_centimeters(140).as_centimeters(), 140);
assert_eq!(Distance::from_centimeters(200).as_meters(), 2);

pub fn from_inches(inches: f32) -> Self

Returns a Distance from the provided inches (25.4 * 10^-3 meters).

Examples

assert_eq!(Distance::from_inches(3.).as_micrometers(), 76_200);
assert_eq!(Distance::from_inches(1.).as_centimeters(), 2);
assert_eq!(Distance::from_inches(60.).as_meters(), 1);

pub const fn as_tens_of_nanometers(&self) -> i32

Returns the number of tens of nanometers (10^-8 meters) this distance represents.

Examples

assert_eq!(Distance::from_tens_of_nanometers(5_900).as_tens_of_nanometers(), 5_900);

pub const fn as_micrometers(&self) -> i32

Returns the number of micrometers (10^-6 meters) this distance represents (the nanometers part is trunked).

Examples

// The .9 is trunked
assert_eq!(Distance::from_tens_of_nanometers(590).as_micrometers(), 5);

pub const fn as_millimeters(&self) -> i32

Returns the number of millimeters (10^-3 meters) this distance represents (the micrometers part is trunked).

Examples

// The .9 is trunked
assert_eq!(Distance::from_micrometers(5_900).as_millimeters(), 5);

pub fn as_millimeters_f32(&self) -> f32

Returns the number of millimeters (10^-3 meters) this distance represents as a f32.

It differs from Self::as_millimeters due to the fact that the millimeters are of type f32 (the micrometers part is not trunked).

Examples

assert_eq!(Distance::from_micrometers(5_900).as_millimeters_f32(), 5.9);
assert_eq!(Distance::from_micrometers(3_210).as_millimeters_f32(), 3.21);

pub const fn as_centimeters(&self) -> i32

Returns the number of centimeters (10^-2 meters) this distance represents (the millimeters part is trunked).

Examples

// The .9 is trunked
assert_eq!(Distance::from_millimeters(59).as_centimeters(), 5);

pub const fn as_meters(&self) -> i32

Returns the number of meters this distance represents (the decimeters part is trunked).

Examples

// The .42 is trunked
assert_eq!(Distance::from_centimeters(142).as_meters(), 1);