Struct firmware_core::utils::measurement::duration::SmallDuration

pub struct SmallDuration { /* private fields */ }

A duration of time with a 10 nanoseconds sensitivity and a range of values that goes up to 10*2^32ns (which is almost 43 seconds).

Implementations

impl SmallDuration

pub const SECOND: Self = _

A SmallDuration with a duration 1 second.

pub const MILLI_SECOND: Self = _

A SmallDuration with a duration 1 millisecond.

pub const MICRO_SECOND: Self = _

A SmallDuration with a duration 1 microsecond.

pub const ZERO: Self = _

A SmallDuration with a duration 0 seconds.

pub const MAX_VALUE: Self = _

A SmallDuration with the highest value storable in this struct (10*2^32 nanoseconds, which is almost 43 seconds).

pub const fn from_tens_of_nanos(tens_of_nanos: u32) -> Self

Returns a SmallDuration with a duration of the provided tens of nanoseconds (10^-8 seconds).

Examples

assert_eq!(SmallDuration::from_tens_of_nanos(1).as_nanos(), 10);
assert_eq!(SmallDuration::from_tens_of_nanos(100).as_nanos(), 1_000);
assert_eq!(SmallDuration::from_tens_of_nanos(100).as_tens_of_nanos(), 100);

pub const fn from_micros(micros: u32) -> Self

Returns a SmallDuration with a duration of the provided microseconds (10^-6 seconds).

Examples

assert_eq!(SmallDuration::from_micros(1).as_nanos(), 1_000);
assert_eq!(SmallDuration::from_micros(1_000_000).as_seconds(), 1);
assert_eq!(SmallDuration::from_micros(100).as_micros(), 100);

pub const fn from_millis(millis: u16) -> Self

Returns a SmallDuration with a duration of the provided milliseconds (10^-3 seconds).

Examples

assert_eq!(SmallDuration::from_millis(1).as_nanos(), 1_000_000);
assert_eq!(SmallDuration::from_millis(1_000).as_seconds(), 1);
assert_eq!(SmallDuration::from_millis(100).as_millis(), 100);

pub const fn from_seconds(seconds: u8) -> Self

Returns a SmallDuration with a duration of the provided seconds.

Keep in mind the max value this struct can represent.

Examples

assert_eq!(SmallDuration::from_seconds(1).as_nanos(), 1_000_000_000);
assert_eq!(SmallDuration::from_seconds(30).as_millis(), 30_000);
assert_eq!(SmallDuration::from_seconds(20).as_seconds(), 20);

pub fn from_seconds_f32(seconds: f32) -> Self

Returns a SmallDuration with a duration of the provided seconds.

It differs from Self::from_seconds due to the fact that the seconds are of type f32 (the subsecond part is not trunked).

Keep in mind the max value this struct can represent.

Examples

assert_eq!(SmallDuration::from_seconds_f32(1.41).as_nanos(), 1_410_000_000);
assert_eq!(SmallDuration::from_seconds_f32(30.2).as_millis(), 30_200);
assert_eq!(SmallDuration::from_seconds_f32(20.202).as_seconds_f32(), 20.202);

pub const fn as_nanos(&self) -> u64

Returns the number of nanoseconds (10^-9 seconds) this duration represents.

pub const fn as_tens_of_nanos(&self) -> u32

Returns the number of tens of nanoseconds (10^-8 seconds) this duration represents.

pub const fn as_micros(&self) -> u32

Returns the number of microseconds (10^-6 seconds) this duration represents (the nanoseconds part is trunked).

Examples

// The .9 is trunked
assert_eq!(SmallDuration::from_tens_of_nanos(590).as_micros(), 5);

pub const fn as_millis(&self) -> u32

Returns the number of milliseconds (10^-3 seconds) this duration represents (the microseconds part is trunked).

Examples

// The .8 is trunked
assert_eq!(SmallDuration::from_micros(5800).as_millis(), 5);

pub const fn as_seconds(&self) -> u32

Returns the number of seconds this duration represents (the milliseconds part is trunked).

Examples

// The .9 is trunked
assert_eq!(SmallDuration::from_millis(5900).as_seconds(), 5);

pub fn as_seconds_f32(&self) -> f32

Returns the number of seconds this duration represents.

It differs from Self::as_seconds due to the fact that the seconds are of type f32 (the subsecond part is not trunked).

Examples

assert_eq!(SmallDuration::from_millis(5_902).as_seconds_f32(), 5.902);
assert_eq!(SmallDuration::from_millis(3_141).as_seconds_f32(), 3.141);

pub fn as_millis_f32(&self) -> f32

Returns the number of milliseconds this duration represents.

It differs from Self::as_millis due to the fact that the milliseconds are of type f32 (the submillisecond part is not trunked).

Examples

assert_eq!(SmallDuration::from_micros(5_902).as_millis_f32(), 5.902);
assert_eq!(SmallDuration::from_micros(3_141).as_millis_f32(), 3.141);