firmware_core::printer
Module communication

This module contains everything related to making the 3D printer able to communicate with the external world. Communication methods may include:

WiFi

USB (not supported yet)

Bluetooth (not supported yet)

The data exchange with the external world is used to monitor the printer’s status and execute commands (like starting the printing of a 3D model).

Modules

communicator

This module handles the communication between the microcontroller and the external world.

http

This module provides functionalities for handling HTTP communication within the firmware, including requests, resource management and ways to communicate to the components thread.

ota

This module provides functionality for Over-The-Air (OTA) updates for the firmware.

security

This module provides security functionality for communications with external systems. It ensures that HTTP requests are validated through password protection, brute-force prevention mechanisms…

Structs

Communication

A struct for managing communication with external entities.

CommunicationConfig

Configuration settings for the communication module.

MultiThreadCommunication

Manages multi-threaded communication for the 3D printer.

Enums

CreationError

Errors that may occur during the creation of the Communication struct.

SendablePeripherals

Represents peripherals that can be sent between threads.

TickError

Errors that may occur during the tick operation of the Communication struct.