Trait firmware_core::printer::components::drivers::spi_flash_memory::FlashMemoryChip

pub trait FlashMemoryChip {
    const LUNS_PER_DEVICE: u32;
    const PLANES_PER_LUN: u32;
    const BLOCKS_PER_PLANE: u32;
    const PAGES_PER_BLOCK: u32;
    const PAGE_SIZE: u32;
    const PAGE_ECC_SIZE: u32;
    const SPI_MODE: Mode;
    const MAX_CLOCK_FREQUENCY: Frequency;
    const MANUFACTURER_ID: u8;
    const DEVICE_ID: u8;

    // Required methods
    fn initialize<Spi: SpiDevice<u8>>(
        spi_flash_memory: &mut SpiFlashMemory<Self, Spi>
    ) -> Result<(), <Spi as ErrorType>::Error>
       where Self: Sized;
    fn contains_bad_block_mark<Spi: SpiDevice<u8>>(
        block_index: u16,
        spi_flash_memory: &mut SpiFlashMemory<Self, Spi>
    ) -> Result<bool, <Spi as ErrorType>::Error>
       where Self: Sized;
}

A type that represents a flash memory chip.

Required Associated Constants

const LUNS_PER_DEVICE: u32

Number of LUNs contained in the chip.

const PLANES_PER_LUN: u32

Number of planes contained in a LUN of the chip.

const BLOCKS_PER_PLANE: u32

Number of blocks contained in a plane of the chip.

const PAGES_PER_BLOCK: u32

Number of pages contained in a block of the chip.

const PAGE_SIZE: u32

Size of the data area of a page.

const PAGE_ECC_SIZE: u32

Size of the ECC area of a page.

const SPI_MODE: Mode

Preferred SPI mode.

const MAX_CLOCK_FREQUENCY: Frequency

Max supported frequency of the SPI clock.

const MANUFACTURER_ID: u8

const DEVICE_ID: u8

Required Methods

fn initialize<Spi: SpiDevice<u8>>( spi_flash_memory: &mut SpiFlashMemory<Self, Spi> ) -> Result<(), <Spi as ErrorType>::Error>where Self: Sized,

fn contains_bad_block_mark<Spi: SpiDevice<u8>>( block_index: u16, spi_flash_memory: &mut SpiFlashMemory<Self, Spi> ) -> Result<bool, <Spi as ErrorType>::Error>where Self: Sized,

Check if the block at the provided block_index in the provided spi_flash_memory contains a bad block mark.

Returns Err(...) if there is a problem in reading from the flash memory and it has been impossible to check if the mark is there. Otherwise returns Ok(contains_bad_block_mark).

Object Safety

This trait is not object safe.