Trait firmware_core::printer::components::motion::homing::endstop::Endstop

pub trait Endstop {
    type IsEndReachedError;
    type OnEndReachedError;
    type HomingError: Debug;

    // Required methods
    fn is_end_reached(&self) -> Result<bool, Self::IsEndReachedError>;
    unsafe fn on_end_reached(
        &mut self,
        callback: impl FnMut() + Send + 'static
    ) -> Result<(), Self::OnEndReachedError>;

    // Provided methods
    fn prepare_for_homing(&mut self) -> Result<(), Self::HomingError> { ... }
    fn finish_homing(&mut self) -> Result<(), Self::HomingError> { ... }
}

Type that can be used as an endstop.

Required Associated Types

type IsEndReachedError

type OnEndReachedError

type HomingError: Debug

Required Methods

fn is_end_reached(&self) -> Result<bool, Self::IsEndReachedError>

Check if the endstop is currently triggered or not.

Returns Err(Self::IsEndReachedError) if there has been an error in reading the endstop’s state, otherwise returns Ok(is_currently_triggered).

unsafe fn on_end_reached( &mut self, callback: impl FnMut() + Send + 'static ) -> Result<(), Self::OnEndReachedError>

Execute the provided callback when the endstop is triggered.

Safety

The callback will be called in an ISR context.

Provided Methods

fn prepare_for_homing(&mut self) -> Result<(), Self::HomingError>

Prepare the endstop for homing. Some types of endstop may not require a preparation so there’s a default blank implementation, but others like a probe with the BLTouch sensor do.

fn finish_homing(&mut self) -> Result<(), Self::HomingError>

The endstop finished homing. Some types of endstop may not need to do anything so there’s a default blank implementation, but others like a probe with the BLTouch sensor do.

Object Safety

This trait is not object safe.