firmware_core::printer::components
Module pauser

This module gives the ability to the printer to pause a print and resume it later.

For now this module only stores the pause state in a static AtomicBool, then other parts of this firmware make a decision on what to do based on what is_paused returns.

Functions

is_paused

Returns false if you’ve never called any function in this module (as a default value) or if after calling them (resume and toggle_pause) you end up with false being the current state.

resume

Makes is_paused return false.

toggle_pause

Makes is_paused return the opposite of what it would return now.