firmware_core::utils::math::vectors
Struct VectorN

pub struct VectorN<const N: usize>(/* private fields */);

An N-dimensional vector with components stored as Distance.

Implementations

impl VectorN<2>

pub const fn from_xy(x: Distance, y: Distance) -> Self

Creates a 2D vector from x and y components.

Examples

let x = Distance::from_millimeters(10);
let y = Distance::from_millimeters(15);
let vector2 = Vector2::from_xy(x, y);
assert_eq!(vector2.x(), x);
assert_eq!(vector2.y(), y);

pub fn x(&self) -> Distance

pub fn y(&self) -> Distance

impl VectorN<3>

pub const fn from_xyz(x: Distance, y: Distance, z: Distance) -> Self

Creates a 3D vector from x, y, and z components.

Examples

let x = Distance::from_millimeters(10);
let y = Distance::from_millimeters(15);
let z = Distance::from_millimeters(5);
let vector3 = Vector3::from_xyz(x, y, z);
assert_eq!(vector3.x(), x);
assert_eq!(vector3.y(), y);
assert_eq!(vector3.z(), z);

pub fn x(&self) -> Distance

pub fn y(&self) -> Distance

pub fn z(&self) -> Distance

impl<const N: usize> VectorN<N>

pub const ZERO: Self = _

A constant representing a zero vector.

Examples

assert_eq!(VectorN::<2>::ZERO.length_millimeters(), 0.);
assert_eq!(VectorN::<3>::ZERO.length_millimeters(), 0.);

pub const fn new(components: [Distance; N]) -> Self

pub const fn get_internal_array(&self) -> &[Distance; N]

pub fn length_millimeters_sqr(&self) -> f32

Returns the square of the length of this vector in millimeters.

Examples

let x = Distance::from_millimeters(3);
let y = Distance::from_millimeters(4);
let vector2 = VectorN::<2>::from_xy(x, y);
assert_eq!(vector2.length_millimeters_sqr() as i32, x.as_millimeters().sqr() + y.as_millimeters().sqr());

pub fn length_millimeters(&self) -> f32

Returns the length of this vector in millimeters.

Examples

let x = Distance::from_millimeters(3);
let y = Distance::from_millimeters(4);
let vector2 = VectorN::<2>::from_xy(x, y);
assert_eq!(vector2.length_millimeters(), vector2.length_millimeters_sqr().sqrt());

pub fn normalized(&self) -> Self

pub fn dot(&self, other: &Self) -> Distance

Returns the dot product of this vector with other.